+++ frontend/src/tables/elements/TableHeader.tsx
@@ -1,6 +1,7 @@
 import { useWebAppOptionsStore } from "../../stores/useWebAppOptionsStore";
 import type { createTableStore } from "../TableStoreFactory";
 import CreateRowButton from "./CreateRowButton";
+import ExportButton from "./ExportButton";
 import HeaderColumnsRow from "./HeaderColumnsRow";
 import PaginationControls from "./PaginationControls";
 import TableReloadButton from "./TableReloadButton";
@@ -17,14 +18,15 @@ export default function TableHeader({ useStore }: Props) {
       <tr>
         <th colSpan={100} style={{ backgroundColor: color.c200 }}>
           <div className="d-flex">
-            <CreateRowButton useStore={useStore}></CreateRowButton>
-            <TableTitle useStore={useStore}></TableTitle>
-            <TableReloadButton useStore={useStore}></TableReloadButton>
-            <PaginationControls useStore={useStore}></PaginationControls>
+            <CreateRowButton useStore={useStore} />
+            <TableTitle useStore={useStore} />
+            <TableReloadButton useStore={useStore} />
+            <PaginationControls useStore={useStore} />
+            <ExportButton useStore={useStore} />
           </div>
         </th>
       </tr>
-      <HeaderColumnsRow useStore={useStore}></HeaderColumnsRow>
+      <HeaderColumnsRow useStore={useStore} />
     </thead>
   );
 }